@@ -1,11 +1,11 @@
 <template>
-  <v-card max-width="400">
+  <v-card density="compact" class="text-caption" max-width="400">
     <v-toolbar dense :color="selectedEvent.color">
       <v-btn icon>
         <v-icon>mdi-information</v-icon>
       </v-btn>
 
-      <v-toolbar-title class="white--text"> Información </v-toolbar-title>
+      <v-toolbar-title class="text-white"> Información </v-toolbar-title>
       <v-spacer />
       <v-btn icon @click="dialogSync = false">
         <v-icon>mdi-close</v-icon>
@@ -13,65 +13,55 @@
     </v-toolbar>
 
     <v-list-item>
-      <v-list-item-icon>
+      <template #prepend>
         <v-icon :color="selectedEvent.color + ' lighten-2'"> mdi-book </v-icon>
-      </v-list-item-icon>
-      <v-list-item-content>
-        <v-list-item-title>
-          {{ selectedEvent.title }}
-        </v-list-item-title>
-      </v-list-item-content>
-      <v-list-item-action>
-        <v-btn icon>
-          <v-icon dense> mdi-information </v-icon>
-        </v-btn>
-      </v-list-item-action>
+      </template>
+      <v-list-item-title>
+        {{ selectedEvent.title }}
+      </v-list-item-title>
     </v-list-item>
     <v-divider inset />
-    <v-list-item dense three-line>
-      <v-list-item-icon>
+    <v-list-item density="compact" lines="three">
+      <template #prepend>
         <v-icon :color="selectedEvent.color + ' lighten-2'"> mdi-text </v-icon>
-      </v-list-item-icon>
-      <v-list-item-content>
-        <v-list-item-title style="white-space: pre-line">
-          {{ selectedEvent.description }}
-        </v-list-item-title>
-      </v-list-item-content>
+      </template>
+
+      <v-list-item-title style="white-space: pre-line">
+        {{ selectedEvent.description }}
+      </v-list-item-title>
     </v-list-item>
     <v-divider inset />
-    <v-list-item dense>
-      <v-list-item-icon>
+    <v-list-item density="compact">
+      <template #prepend>
         <v-icon :color="selectedEvent.color + ' lighten-2'">
           mdi-calendar
         </v-icon>
-      </v-list-item-icon>
-      <v-list-item-content>
-        <v-list-item-title>
-          {{ selectedDay }}, {{ selectedEvent.startTime }}-{{
-            selectedEvent.endTime
-          }}
-        </v-list-item-title>
-        <v-list-item-subtitle>{{ selectedEvent.type }}</v-list-item-subtitle>
-      </v-list-item-content>
+      </template>
+
+      <v-list-item-title>
+        {{ selectedDay }}, {{ selectedEvent.startTime }}-{{
+          selectedEvent.endTime
+        }}
+      </v-list-item-title>
+      <v-list-item-subtitle>{{ selectedEvent.type }}</v-list-item-subtitle>
     </v-list-item>
 
     <v-divider inset />
-    <v-list-item dense>
-      <v-list-item-icon>
+    <v-list-item density="compact">
+      <template #prepend>
         <v-icon :color="selectedEvent.color + ' lighten-2 '">
           mdi-map-marker
         </v-icon>
-      </v-list-item-icon>
-      <v-list-item-content>
-        <v-list-item-title>{{ selectedEvent.location }}</v-list-item-title>
-      </v-list-item-content>
+      </template>
+      <v-list-item-title>{{ selectedEvent.location }}</v-list-item-title>
     </v-list-item>
   </v-card>
 </template>
 <script lang="ts">
 import { useVModel } from '@vueuse/core'
-import { defineComponent, PropType, computed, Ref } from 'vue'
+import { defineComponent, type PropType, computed, type Ref } from 'vue'
 import Event from '~/model/Event'
+import { WEEK_DAYS } from '~/constants/weekdays'
 
 export default defineComponent({
   props: {
@@ -88,18 +78,8 @@ export default defineComponent({
   setup(props, { emit }) {
     const dialogSync = useVModel(props, 'dialog', emit)
 
-    const weekdays = [
-      'Sunday',
-      'Monday',
-      'Tuesday',
-      'Wednesday',
-      'Thursday',
-      'Friday',
-      'Saturday',
-    ]
-
     const selectedDay: Ref<string> = computed(
-      () => weekdays[props.selectedEvent.day]
+      () => WEEK_DAYS[props.selectedEvent.day],
     )
 
     const moreInfo = () => {
@@ -108,7 +88,6 @@ export default defineComponent({
 
     return {
       selectedDay,
-      weekdays,
       moreInfo,
       dialogSync,
     }
