@@ -3,22 +3,23 @@
     <v-col cols="12" class="text-h4 text-center">
       Agrega tus horarios favoritos
     </v-col>
-    <v-col id="bm" cols="12" />
+    <v-col ref="bm" cols="12" />
   </v-row>
 </template>
-<script lang="ts">
+<script setup lang="ts">
 import Lottie from 'lottie-web'
+import type { VCol } from 'vuetify/components'
+import Animation from '~/assets/lottie/156-star-blast.json'
 
-export default {
-  name: 'FavoriteBanner',
-  mounted() {
-    Lottie.loadAnimation({
-      container: document.getElementById('bm') as Element,
-      renderer: 'svg',
-      loop: false,
-      autoplay: true,
-      animationData: require('~/assets/lottie/156-star-blast.json'),
-    })
-  },
-}
+const bm = ref<ComponentPublicInstance  | null>(null)
+
+onMounted(() => {
+  Lottie.loadAnimation({
+    container: bm.value?.$el,
+    renderer: 'svg',
+    loop: false,
+    autoplay: true,
+    animationData: Animation,
+  })
+})
 </script>
