@@ -7,7 +7,7 @@
       v-else
       rounded
       class="ma-1"
-      shaped
+      variant="outlined"
       @click="dialogCalendarSync = true"
     >
       Agregar a mi Calendario
@@ -28,13 +28,13 @@
             <template v-if="signInStatus">
               <v-autocomplete
                 v-model="selected"
+                v-model:search-input="search"
                 label="Selecciona tu calendario"
                 :items="calendarList"
                 return-object
                 clearable
                 :rules="[(r) => !!r || 'Requerido']"
-                :search-input.sync="search"
-                item-text="summary"
+                item-title="summary"
               >
                 <template #prepend-item="">
                   <v-list-item
@@ -42,11 +42,9 @@
                     color="primary"
                     @click="addCalendar()"
                   >
-                    <v-list-item-content class="text--primary">
-                      <v-list-item-title class="text primary--text">
-                        Crear mi calendario...
-                      </v-list-item-title>
-                    </v-list-item-content>
+                    <v-list-item-title class="text text-primary">
+                      Crear mi calendario...
+                    </v-list-item-title>
                   </v-list-item>
                 </template>
               </v-autocomplete>
@@ -71,9 +69,9 @@
                   v-model="notifications[index].minutes"
                   type="number"
                   :rules="[(a) => a > 0 || 'No permitido']"
-                  outlined
-                  dense
-                  :value="notification.minutes"
+                  variant="outlined"
+                  density="compact"
+                  :model-value="notification.minutes"
                   prepend-icon="mdi-bell"
                   :items="Array.from({ length: 60 }, (x, i) => i)"
                   suffix="minutos"
@@ -92,13 +90,13 @@
                   key="selected"
                   v-model="defaultNotification.minutes"
                   :rules="[(a) => a > 0 || 'No permitido']"
-                  :value="defaultNotification.minutes"
+                  :model-value="defaultNotification.minutes"
                   prepend-icon="mdi-bell"
                   type="number"
                   :items="Array.from({ length: 60 }, (x, i) => i)"
                   suffix="minutos"
-                  outlined
-                  dense
+                  variant="outlined"
+                  density="compact"
                 >
                   <template #append-outer>
                     <v-icon color="success" @click="addNotification()">
@@ -111,23 +109,27 @@
           </v-form>
         </v-card-text>
         <v-card-actions v-if="signInStatus">
-          <v-btn :loading="progress > 0" text @click="exportEventToGCalendar">
+          <v-btn
+            :loading="progress > 0"
+            variant="text"
+            @click="exportEventToGCalendar"
+          >
             Exportar
             <template #loader>
               <v-progress-linear
                 color="secondary"
                 striped
-                :value="(progress / events.length) * 100"
+                :model-value="(progress / events.length) * 100"
                 :height="30"
               >
                 <template #default="{ value }">
-                  <strong class="white--text">{{ Math.ceil(value) }}%</strong>
+                  <strong class="text-white">{{ Math.ceil(value) }}%</strong>
                 </template>
               </v-progress-linear>
             </template>
           </v-btn>
           <v-spacer />
-          <v-btn text color="success" @click="handleSignOutClick()">
+          <v-btn variant="text" color="success" @click="handleSignOutClick()">
             Cerrar Sesión
           </v-btn>
         </v-card-actions>
@@ -137,15 +139,20 @@
 </template>
 
 <script lang="ts">
-import { ref, computed, defineComponent, PropType, toRefs, watch } from 'vue'
+import {
+  ref,
+  computed,
+  defineComponent,
+  type PropType,
+  toRefs,
+  watch,
+} from 'vue'
 import { DateTime } from 'luxon'
-import { useContext } from '@nuxtjs/composition-api'
 import { v4 } from 'uuid'
 import { colors } from '~/utils/core'
 import CreateGoogleCalendar from '~/components/CreateGoogleCalendar.vue'
 import GoogleSignIn from '~/components/GoogleSignIn.vue'
 import Event from '~/model/Event'
-import { VForm } from '~/types'
 
 export default defineComponent({
   name: 'GoogleAuth',
@@ -192,10 +199,10 @@ export default defineComponent({
     const dialog = ref(false)
     const dayNames = ['Do', 'Lu', 'Ma', 'Mi', 'Ju', 'Vi', 'Sa']
     const dateStart = computed(() =>
-      DateTime.fromISO(props.startDate).toFormat('yyyy-MM-dd')
+      DateTime.fromISO(props.startDate).toFormat('yyyy-MM-dd'),
     )
     const dateEnd = computed(() =>
-      DateTime.fromISO(props.endDate).toFormat('yyyy-MM-dd')
+      DateTime.fromISO(props.endDate).toFormat('yyyy-MM-dd'),
     )
 
     const progress = ref(0)
@@ -231,7 +238,7 @@ export default defineComponent({
       window.gapi.load('client:auth2', initClient)
     }
 
-    const config = useContext().$config
+    const config = useNuxtApp().$config.public
     const initClient = async () => {
       try {
         loading.value = true
@@ -276,7 +283,7 @@ export default defineComponent({
       console.log(index)
       notifications.value = notifications.value.filter(
         (notification: { method: string; minutes: number }) =>
-          notification !== index
+          notification !== index,
       )
       defaultNotification = {
         method: 'popup',
@@ -316,7 +323,7 @@ export default defineComponent({
       }
     }
 
-    const form = ref<VForm>()
+    const form = ref<any>()
 
     const exportEventToGCalendar = async () => {
       if (!form.value?.validate()) {
@@ -352,7 +359,7 @@ export default defineComponent({
           start: {
             dateTime: DateTime.fromFormat(
               dateStart.value + ' ' + event.startTime,
-              format
+              format,
             )
               .set({ weekday: event.day })
               .toISO(),
@@ -361,7 +368,7 @@ export default defineComponent({
           end: {
             dateTime: DateTime.fromFormat(
               dateStart.value + ' ' + event.endTime,
-              format
+              format,
             )
               .set({ weekday: event.day })
               .toISO(),
@@ -438,6 +445,14 @@ export default defineComponent({
       dayNames,
       day,
       signInStatus,
+      progress,
+      defaultNotification,
+      dialog,
+      loading,
+      selected,
+      calendarList,
+      search,
+      calendarItem,
     }
   },
   head() {
