@@ -1,16 +1,17 @@
 <template>
   <v-sheet
     :color="event.color"
-    class="white--text text-wrap pa-1 ma-0"
-    style="max-width: inherit"
+    class="text-white text-wrap pl-2 pa-1 ma-0 border-1"
+    width="100%"
   >
     <template v-if="!event.code">
-      <span class="font-weight-black">
+      <v-icon start size="small">mdi-information</v-icon>
+      <span class="font-weight-bold">
         {{ event.title }}
       </span>
     </template>
     <template v-else>
-      <span class="font-weight-black">
+      <span>
         {{ event.code }}
         {{ event.section }}
         ({{ event.type }})
@@ -18,25 +19,24 @@
       <br />
       {{ event.name }}
     </template>
-    <br />
-    <v-fade-transition>
-      <v-overlay v-if="hover" z-index="0" absolute :color="event.color">
-        <v-icon>mdi-information</v-icon>
-      </v-overlay>
-    </v-fade-transition>
   </v-sheet>
 </template>
 <script lang="ts">
-export default {
+import { defineComponent } from 'vue'
+export default defineComponent({
   name: 'ScheduleEventInfo',
   props: {
-    event: {},
-    hover: {},
+    event: {
+      type: Object,
+      required: true,
+    },
   },
-}
+})
 </script>
-<style lang="sass">
-
-div.v-calendar-daily__day:nth-child(n) > div:nth-child(17)
-  margin: 0
+<style lang="scss">
+.border-1 {
+  border-radius: 5px;
+  box-shadow: 0px 0px 8px 0px rgba(0, 0, 0, 0.2);
+  border: 1px solid #e0e0e0;
+}
 </style>
