@@ -1,10 +1,14 @@
 <template>
   <v-card>
-    <v-card-title>
-      <span class="title font-weight-bold">Compartir</span>
+    <v-card-title class="d-flex">
+      <span class="text-h6 font-weight-bold">Compartir</span>
       <v-spacer />
-      <v-btn class="mx-0" icon @click="close">
-        <v-icon>mdi-close-circle-outline</v-icon>
+      <v-btn
+        class="mx-0"
+        variant="text"
+        icon="mdi-close-circle-outline"
+        @click="close"
+      >
       </v-btn>
     </v-card-title>
     <v-list>
@@ -12,9 +16,9 @@
         :href="'https://www.facebook.com/sharer/sharer.php?u=' + link"
         target="_blank"
       >
-        <v-list-item-action>
+        <template #prepend>
           <v-icon color="indigo"> mdi-facebook </v-icon>
-        </v-list-item-action>
+        </template>
         <v-card-title>Facebook</v-card-title>
       </v-list-item>
       <v-list-item
@@ -25,9 +29,9 @@
         "
         target="_blank"
       >
-        <v-list-item-action>
+        <template #prepend>
           <v-icon color="cyan"> mdi-twitter </v-icon>
-        </v-list-item-action>
+        </template>
         <v-card-title>Twitter</v-card-title>
       </v-list-item>
       <v-list-item
@@ -35,9 +39,9 @@
         target="_blank"
         data-action="share/whatsapp/share"
       >
-        <v-list-item-action>
+        <template #prepend>
           <v-icon color="success"> mdi-whatsapp </v-icon>
-        </v-list-item-action>
+        </template>
         <v-card-title>Whatsapp</v-card-title>
       </v-list-item>
       <v-list-item
@@ -45,9 +49,9 @@
         target="_blank"
         data-action="share/telegram/share"
       >
-        <v-list-item-action>
+        <template #prepend>
           <v-icon color="#54a9eb"> mdi-telegram </v-icon>
-        </v-list-item-action>
+        </template>
         <v-card-title>Telegram</v-card-title>
       </v-list-item>
     </v-list>
@@ -57,57 +61,66 @@
       :label="copied ? 'Enlace copiado' : 'Click to copiar el link'"
       class="pa-4"
       readonly
-      :value="link"
+      :model-value="link"
       @click="copy"
     />
   </v-card>
 </template>
-<script lang="js">
+<script lang="ts">
+import { defineComponent } from 'vue'
+import type { IScheduleGenerate } from '~/interfaces/schedule'
 
-export default {
+export default defineComponent({
   name: 'ScheduleShare',
   props: {
-    schedule: { type: Object },
+    schedule: {
+      type: Object as PropType<IScheduleGenerate>,
+      required: true,
+    },
     path: { type: String, default: '/subject' },
     dialog: { type: Boolean },
-    postId: { type: Number, default: 1 }
+    postId: { type: Number, default: 1 },
   },
+  emits: ['update:dialog'],
   data: () => ({
     copied: false,
-    dialogSync: true
+    dialogSync: true,
   }),
   computed: {
-    link () {
+    link() {
       return location.origin + this.path + '?q=' + this.query
     },
-    query () {
+    query() {
       return btoa(this.schedule.scheduleSubjectIds.join(','))
-    }
+    },
   },
   watch: {
-    dialog (newValue) {
+    dialog(newValue) {
       this.dialogSync = newValue
     },
-    dialogSync (newValue) {
+    dialogSync(newValue) {
       this.$emit('update:dialog', newValue)
-    }
+    },
   },
   methods: {
-    close () {
+    close() {
       this.dialogSync = false
     },
-    copy () {
+    copy() {
       const copyText = document.querySelector('#link')
       if (copyText) {
         copyText.select()
-        navigator.clipboard.writeText(this.link).then(() => {
-          /* clipboard successfully set */
-          this.copied = true
-        }, function () {
-          /* clipboard write failed */
-        })
+        navigator.clipboard.writeText(this.link).then(
+          () => {
+            /* clipboard successfully set */
+            this.copied = true
+          },
+          function () {
+            /* clipboard write failed */
+          },
+        )
       }
-    }
-  }
-}
+    },
+  },
+})
 </script>
