@@ -1,14 +1,14 @@
 <template>
-  <v-list-item three-line dense>
-    <v-list-item-content v-if="hourlyLoad">
+  <v-list-item lines="three" density="compact">
+    <template v-if="hourlyLoad">
       {{ hourlyLoad.name }}
       <v-list-item-subtitle>
         Ult. Actual.: {{ new Date(hourlyLoad.updatedAt).toLocaleString() }}
       </v-list-item-subtitle>
-    </v-list-item-content>
-    <v-list-item-content v-else>
+    </template>
+    <template v-else>
       <v-list-item-subtitle> Sin carga horaria. </v-list-item-subtitle>
-    </v-list-item-content>
+    </template>
   </v-list-item>
 </template>
 <script lang="ts">
