@@ -4,7 +4,7 @@
       <v-checkbox
         v-model="valueSync"
         class="text-caption"
-        dense
+        density="compact"
         :label="section"
         :value="schedule"
         multiple
@@ -13,8 +13,8 @@
   </tr>
 </template>
 <script lang="ts">
-import { defineComponent, PropType, computed } from 'vue'
-import { ISubjectSchedule } from '~/interfaces/subject'
+import { defineComponent, type PropType, computed } from 'vue'
+import type { ISubjectSchedule } from '~/interfaces/subject'
 
 export default defineComponent({
   props: {
@@ -22,17 +22,17 @@ export default defineComponent({
       type: Object as PropType<ISubjectSchedule>,
       required: true,
     },
-    value: {
+    modelValue: {
       type: Array as PropType<ISubjectSchedule[]>,
       default: null,
     },
   },
-  emits: ['input'],
+  emits: ['update:modelValue'],
   setup(props, { emit }) {
     const valueSync = computed({
-      get: () => props.value,
+      get: () => props.modelValue,
       set: (newValue) => {
-        emit('input', newValue)
+        emit('update:modelValue', newValue)
       },
     })
 
