@@ -1,6 +1,4 @@
-import { useContext } from '@nuxtjs/composition-api'
-
 export const useApi = () => {
-  const { $api } = useContext()
+  const { $api } = useNuxtApp()
   return $api
 }
