@@ -1,19 +1,27 @@
 <template>
-  <div>
-    <NuxtChild />
-    <v-dialog v-if="firstEntry" v-model="firstEntry" max-width="600" persistent>
-      <InitialSettings :dialog.sync="firstEntry" />
+  <v-container>
+    <v-dialog
+      v-show="firstEntry"
+      v-model="firstEntry"
+      max-width="600"
+      persistent
+    >
+      <InitialSettings />
     </v-dialog>
     <v-dialog v-model="isNewHourlyLoad" max-width="600">
       <v-card>
-        <v-card-title class="headline">Nueva Carga Horaria</v-card-title>
+        <v-card-title class="text-h5">Nueva Carga Horaria</v-card-title>
         <v-card-text>
           Se ha encontrado una nueva carga horaria. Actualiza las secciones de
           los cursos que ya tenias guardadas si deseas los nuevos horarios.
         </v-card-text>
         <v-card-actions>
           <v-spacer></v-spacer>
-          <v-btn color="green darken-1" text @click="isNewHourlyLoad = false">
+          <v-btn
+            color="green-darken-1"
+            variant="text"
+            @click="isNewHourlyLoad = false"
+          >
             Cerrar
           </v-btn>
         </v-card-actions>
@@ -21,7 +29,7 @@
     </v-dialog>
     <v-dialog v-model="isUpdateHourlyLoad" persistent>
       <v-card>
-        <v-card-title class="headline"
+        <v-card-title class="text-h5"
           >Se ha actualizado la Carga Horaria</v-card-title
         >
         <v-card-text>
@@ -31,8 +39,8 @@
         <v-card-actions>
           <v-spacer></v-spacer>
           <v-btn
-            color="green darken-1"
-            text
+            color="green-darken-1"
+            variant="text"
             @click="isUpdateHourlyLoad = false"
           >
             Cerrar
@@ -40,24 +48,33 @@
         </v-card-actions>
       </v-card>
     </v-dialog>
-  </div>
+    <NuxtPage />
+  </v-container>
 </template>
 
 <script lang="ts">
-import { useFetch, useRouter } from '@nuxtjs/composition-api'
 import { storeToRefs } from 'pinia'
 import { defineComponent, watch, onMounted } from 'vue'
 import InitialSettings from '~/components/setting/Initial.vue'
 import { useUserConfigStore } from '~/stores/user-config'
 import { useUserEventsStore } from '~/stores/user-events'
 
+
 export default defineComponent({
   name: 'Generator',
   components: {
     InitialSettings,
   },
-  layout: 'app',
   setup() {
+    definePageMeta({
+      layout: 'app',
+    })
+
+    useSeoMeta({
+      title: 'Generador de Horarios',
+      description: 'Genera tu horario de clases de manera sencilla',
+    })
+
     const configStore = useUserConfigStore()
     const eventsStore = useUserEventsStore()
     const router = useRouter()
@@ -65,18 +82,12 @@ export default defineComponent({
     const { firstEntry, isNewHourlyLoad, isUpdateHourlyLoad } =
       storeToRefs(configStore)
 
-    watch(firstEntry, (newValue, oldValue) => {
+    watch(firstEntry, async (newValue, oldValue) => {
       if (oldValue && !newValue) {
-        router.push('/generator/subjects')
+        await router.push('/generator/subjects')
       }
     })
-    const { fetch, fetchState } = useFetch(async () => {
-      const store = useUserConfigStore()
-      await store.fetchFirstEntry()
-      await store.fetchFaculty()
-      await store.fetchSpeciality()
-      await store.fetchHourlyLoad()
-    })
+
 
     onMounted(async () => {
       await configStore.fetchSubjects()
@@ -90,8 +101,6 @@ export default defineComponent({
       firstEntry,
       isNewHourlyLoad,
       isUpdateHourlyLoad,
-      fetch,
-      fetchState,
     }
   },
   head: {
