@@ -1,16 +1,16 @@
 <template>
   <schedules-presentation
+    v-model:dialog="openMySchedules"
     color="blue"
     title="Generados"
     empty-message="Usted no tiene horarios generados"
     :schedules="schedules"
-    :dialog.sync="openMySchedules"
     path="/skd"
   >
     <template #top-items-right>
       <v-toolbar-title> Generados </v-toolbar-title>
 
-      <v-btn plain class="ml-2" outlined fab small>
+      <v-btn class="ml-2" variant="outlined" size="small">
         {{ schedules.length }}
       </v-btn>
     </template>
@@ -18,7 +18,7 @@
       <schedule-favorite-add
         :favorites-schedules="myFavoritesSchedules"
         :schedule="item"
-        @update:favoritesSchedules="updateFavoritesSchedules"
+        @update:favorites-schedules="updateFavoritesSchedules"
       />
     </template>
     <template #subtitle-items>
@@ -26,8 +26,8 @@
         v-model.number="crossingSubjects"
         label="Cantidad de cruces"
         hide-details
-        outlined
-        dense
+        variant="outlined"
+        density="compact"
         max="5"
         min="0"
         class="shrink"
@@ -36,20 +36,25 @@
 
       <v-btn
         color="success"
-        dark
+        theme="dark"
         rounded
-        shaped
+        variant="outlined"
         class="ma-1"
         @click="generateAllUserSchedules"
       >
         <v-icon>mdi-update</v-icon>
         Generar
       </v-btn>
-      <v-snackbar v-model="succces" color="success" app timeout="3000">
+      <v-snackbar v-model="succces" color="success" timeout="3000">
         <v-icon> mdi-check </v-icon>
         Horarios generados correctamente!
-        <template #action="{ attrs }">
-          <v-btn text small icon v-bind="attrs" @click="succces = false">
+        <template #actions>
+          <v-btn
+            variant="text"
+            size="small"
+            icon
+            @click="succces = false"
+          >
             <v-icon> mdi-close </v-icon>
           </v-btn>
         </template>
@@ -61,7 +66,7 @@
         <v-alert prominent type="error">
           <v-row align="center">
             <v-col class="grow">
-              Lo sentimos, no hemos encontrados horarios :(
+              Lo sentimos, no hemos encontrados horarios para usted.
             </v-col>
           </v-row>
         </v-alert>
@@ -80,7 +85,8 @@ import ScheduleFavoriteAdd from '~/components/ScheduleFavoriteAdd.vue'
 import OccurrencesList from '~/components/OccurrencesList.vue'
 import { useUserConfigStore } from '~/stores/user-config'
 import { useUserEventsStore } from '~/stores/user-events'
-import { ISchedule } from '~/interfaces/schedule'
+import type { IScheduleGenerate } from '~/interfaces/schedule'
+import type { IOccurrence } from '~/interfaces/ocurrences'
 
 export default defineComponent({
   components: {
@@ -88,11 +94,10 @@ export default defineComponent({
     SchedulesPresentation,
     OccurrencesList,
   },
-  layout: 'app',
   setup() {
     const configStore = useUserConfigStore()
     const eventsStore = useUserEventsStore()
-    const occurrences = ref<any[]>([])
+    const occurrences = ref<IOccurrence[]>([])
     const openMySchedules = ref(false)
     const succces = ref(false)
 
@@ -108,11 +113,11 @@ export default defineComponent({
       configStore.updateCrossings(crossings)
     }
 
-    const updateFavoritesSchedules = (favoritesSchedules: ISchedule[]) => {
+    const updateFavoritesSchedules = (favoritesSchedules: IScheduleGenerate[]) => {
       configStore.updateFavoritesSchedules(favoritesSchedules)
     }
 
-    const updateSchedules = (schedules: ISchedule[]) => {
+    const updateSchedules = (schedules: IScheduleGenerate[]) => {
       configStore.updateSchedules(schedules)
     }
 
@@ -123,7 +128,7 @@ export default defineComponent({
         myEvents.value,
         {
           crossingSubjects: crossingSubjects.value,
-        }
+        },
       )
       updateSchedules(combinations)
       occurrences.value = occurrencesData
