@@ -4,10 +4,10 @@
       <v-toolbar flat>
         <v-btn
           v-if="schedules.length > 0"
-          outlined
+          variant="outlined"
           @click="addFavoriteCurrentSchedule"
         >
-          <v-icon :color="isFavorite(schedules[0]) >= 0 ? 'yellow' : null">
+          <v-icon :color="isFavorite(schedules[0]) >= 0 ? 'yellow' : undefined">
             mdi-star
           </v-icon>
           <span v-if="isFavorite(schedules[0]) >= 0">
@@ -27,21 +27,29 @@
 
 <script lang="ts">
 import { defineComponent, ref, computed, onMounted } from 'vue'
-import { useAsync, useRoute } from '@nuxtjs/composition-api'
 import { getSchedules } from '~/utils/core'
 import ScheduleViewer from '~/components/ScheduleViewer.vue'
 import { useUserConfigStore } from '~/stores/user-config'
 import { useApi } from '~/composables/api'
+import type { IScheduleGenerate } from '~/interfaces/schedule'
+
+definePageMeta({
+  layout: 'app',
+})
 
 export default defineComponent({
   components: {
     ScheduleViewer,
   },
-  layout: 'app',
   setup() {
+    useSeoMeta({
+      title: 'Horario compartido',
+      description: 'Comparte tu horario a tus amigos! ',
+    })
+
     const $api = useApi()
     const query = ref('')
-    const schedules = ref<any[]>([])
+    const schedules = ref<IScheduleGenerate[]>([])
     const loading = ref(false)
     const courses = ref([])
 
@@ -50,30 +58,29 @@ export default defineComponent({
     const myFavoritesSchedules = computed(() => store.favoritesSchedules)
     const route = useRoute()
 
-    const data = useAsync(async () => {
-      const query: any = route.value.query
-      const result = Buffer.from(query.q, 'base64').toString()
-      const { data: scheduleSubjects } = await $api.scheduleSubject.getAllByIds(
-        result.split(',').map(Number)
+    const { data } = useAsyncData(async () => {
+      const query: any = route.query
+      const result = decodeBase64(query.q)
+      const scheduleSubjects = await $api.scheduleSubject.getAllByIds(
+        result.split(',').map(Number),
       )
-      const schedulesId = scheduleSubjects.map((ss: any) => ss.schedule.id)
-      const { data: sessions } = await $api.classSessions.findScheduleIds(
-        schedulesId
-      )
+      const schedulesId = scheduleSubjects.map((ss) => ss.schedule.id)
+      const sessions = await $api.classSessions.findScheduleIds(schedulesId)
       return { scheduleSubjects, sessions }
     })
 
-    const scheduleSubjects = computed(() => data.value?.scheduleSubjects)
-    const sessions = computed(() => data.value?.sessions)
+    const scheduleSubjects = computed(() => data.value?.scheduleSubjects || [])
+    const sessions = computed(() => data.value?.sessions || [])
 
-    const deleteFavoriteScheduleById = (favorites: any) =>
-      store.deleteFavoriteScheduleById(favorites)
+    const deleteFavoriteScheduleById = (favorites: IScheduleGenerate) =>
+      store.deleteFavoriteScheduleById(favorites.id)
 
-    const saveNewFavoriteSchedule = (favorites: any) =>
+    const saveNewFavoriteSchedule = (favorites: IScheduleGenerate) =>
       store.saveNewFavoriteSchedule(favorites)
 
     const subjects = computed(() => {
-      return scheduleSubjects.value.map((sb: any) => ({
+      const _sessions = sessions.value
+      return scheduleSubjects.value.map((sb) => ({
         ...sb.subject,
         schedules: [
           {
@@ -81,9 +88,8 @@ export default defineComponent({
             scheduleSubject: {
               id: sb.id,
             },
-            sessions: sessions.value.filter(
-              (s: any) => s.schedule.id === sb.schedule.id
-            ),
+            sessions: _sessions.filter((s) => s.schedule.id === sb.schedule.id),
+            subject: sb.subject,
           },
         ],
       }))
@@ -111,9 +117,9 @@ export default defineComponent({
       }
     }
 
-    const isFavorite = (schedule: any) => {
+    const isFavorite = (schedule: IScheduleGenerate) => {
       return myFavoritesSchedules.value.findIndex(
-        (x: { id: any }) => x.id === schedule.id
+        (x: { id: any }) => x.id === schedule.id,
       )
     }
 
@@ -127,18 +133,6 @@ export default defineComponent({
       addFavoriteCurrentSchedule,
       isFavorite,
       myFavoritesSchedules,
-    }
-  },
-  head() {
-    return {
-      title: 'Horario compartido',
-      meta: [
-        {
-          hid: 'description',
-          name: 'description',
-          content: 'Comparte tu horario a tus amigos! ',
-        },
-      ],
     }
   },
 })
