@@ -1,40 +1,25 @@
-import { Plugin } from '@nuxt/types'
-import axios from 'axios'
 import repositories from '~/repositories'
-export type Repositories = ReturnType<typeof repositories>
 
-interface RepositoryPlugin {
-  $api: Repositories
-}
-
-declare module 'vue/types/vue' {
-  // this.$myInjectedFunction inside Vue components
-  interface Vue extends RepositoryPlugin {}
-}
-
-declare module '@nuxt/types' {
-  // nuxtContext.app.$myInjectedFunction inside asyncData, fetch, plugins, middleware, nuxtServerInit
-  interface NuxtAppOptions extends RepositoryPlugin {}
-  // nuxtContext.$myInjectedFunction
-  interface Context extends RepositoryPlugin {}
-}
-
-declare module 'vuex/types/index' {
-  // this.$myInjectedFunction inside Vuex stores
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  interface Store<S> {
-    $api: RepositoryPlugin
+declare module '#app' {
+  interface NuxtApp {
+    $api: ReturnType<typeof repositories>
   }
 }
 
-const api: Plugin = ({ $config }, inject) => {
+declare module 'vue' {
+  interface ComponentCustomProperties {
+    $api: ReturnType<typeof repositories>
+  }
+}
+
+const api = defineNuxtPlugin((nuxtApp) => {
   // Create a custom axios instance
   // Create a custom axios instance
-  const api = axios.create({
-    baseURL: $config.baseURL + '/api',
+  const api = $fetch.create({
+    baseURL: '/api',
   })
 
-  inject('api', repositories(api))
-}
+  nuxtApp.provide('api', repositories(api))
+})
 
 export default api
