@@ -1,15 +1,18 @@
-import { AxiosInstance } from 'axios'
-import { IPage } from '~/interfaces/page'
-import { ISubject } from '~/interfaces/subject'
+import type { $Fetch } from 'nitropack'
+import type { IPage } from '~/interfaces/page'
+import type { ISubject } from '~/interfaces/subject'
 
 const PATH_SUBJECTS = 'subjects'
-export default ($axios: AxiosInstance) => ({
+export class CourseRepository {
+  constructor(private $fetch: $Fetch) {}
+
   findBySearch(search: string, speciality: number, hourlyLoad: number) {
-    return $axios.get<IPage<ISubject>>(PATH_SUBJECTS + '?search=' + search, {
+    return this.$fetch<IPage<ISubject>>(PATH_SUBJECTS + '?search=' + search, {
+      method: 'GET',
       params: {
         speciality,
         hourlyLoad,
       },
     })
-  },
-})
+  }
+}
