@@ -1,9 +1,11 @@
-import { AxiosInstance } from 'axios'
-import { IOrganization } from '~/interfaces/organization'
+import type { $Fetch } from 'nitropack'
+import type { IOrganization } from '~/interfaces/organization'
 
 const PATH_SUBJECTS = 'faculties'
-export default ($axios: AxiosInstance) => ({
+export class FacultyRepository {
+  constructor(private $fetch: $Fetch) {}
+
   getAll() {
-    return $axios.get<IOrganization[]>(PATH_SUBJECTS)
-  },
-})
+    return this.$fetch<IOrganization[]>(PATH_SUBJECTS)
+  }
+}
