@@ -1,13 +1,16 @@
-import { AxiosInstance } from 'axios'
-import { IHourlyLoad } from '~/interfaces/houly-load'
+import type { $Fetch } from 'nitropack'
+import type { IHourlyLoad } from '~/interfaces/houly-load'
 
 const PATH_SUBJECTS = 'hourlyLoads'
-export default ($axios: AxiosInstance) => ({
+export class HourlyLoadRepository {
+  constructor(private $fetch: $Fetch) {}
+
   getLatestByFaculty(facultyId: any) {
-    return $axios.get<IHourlyLoad>(PATH_SUBJECTS + '/latest', {
+    return this.$fetch<IHourlyLoad>(PATH_SUBJECTS + '/latest', {
+      method: 'GET',
       params: {
         faculty: facultyId,
       },
     })
-  },
-})
+  }
+}
