@@ -1,29 +1,40 @@
-import { AxiosInstance } from 'axios'
-import { IScheduleSubject } from '~/interfaces/schedule-subject'
+import type { $Fetch } from 'nitropack'
+import type {
+  IScheduleSubject,
+  IScheduleSubjectSessionDetail,
+} from '~/interfaces/schedule-subject'
 
 const PATH_SCHEDULE_SUBJECTS = 'scheduleSubjects'
-export default ($axios: AxiosInstance) => ({
+export class ScheduleSubjectRepository {
+  constructor(private $fetch: $Fetch) {}
+
   findBySubjectIdAndHourlyLoadId(subject: number, hourlyLoad: number) {
-    return $axios.get<IScheduleSubject[]>(PATH_SCHEDULE_SUBJECTS, {
+    return this.$fetch<IScheduleSubject[]>(PATH_SCHEDULE_SUBJECTS, {
       params: {
         subject,
         hourlyLoad,
       },
     })
-  },
+  }
+
   getAllByIds(ids: Array<number>) {
-    return $axios.get(PATH_SCHEDULE_SUBJECTS, {
-      params: {
-        ids: ids.join(','),
+    return this.$fetch<IScheduleSubjectSessionDetail[]>(
+      PATH_SCHEDULE_SUBJECTS,
+      {
+        params: {
+          ids: ids.join(','),
+        },
       },
-    })
-  },
+    )
+  }
+
   findBySearch(search: string, speciality: string, hourlyLoad: string) {
-    return $axios.get(PATH_SCHEDULE_SUBJECTS + '?search=' + search, {
+    return this.$fetch(PATH_SCHEDULE_SUBJECTS, {
       params: {
         speciality,
         hourlyLoad,
+        search: search,
       },
     })
-  },
-})
+  }
+}
