@@ -1,12 +1,14 @@
-import { AxiosInstance } from 'axios'
-import { IOrganization } from '~/interfaces/organization'
+import type { $Fetch } from 'nitropack'
+import type { IOrganization } from '~/interfaces/organization'
 const PATH_SUBJECTS = 'specialities'
-export default ($axios: AxiosInstance) => ({
-  getAllByFaculty(facultyId: any) {
-    return $axios.get<IOrganization[]>(PATH_SUBJECTS, {
+export class SpecialityRepository {
+  constructor(private $fetch: $Fetch) {}
+
+  getAllByFaculty(facultyId: number) {
+    return this.$fetch<IOrganization[]>(PATH_SUBJECTS, {
       params: {
         faculty: facultyId,
       },
     })
-  },
-})
+  }
+}
