@@ -3,14 +3,14 @@ import { ref } from 'vue'
 
 export interface SnackbarOptions {
   content: string
-  timeout?: number | null
-  color?: string | null
+  timeout?: number
+  color?: string
 }
 
 export const useGlobalSnackbarStore = defineStore('global/snackbar', () => {
   const content = ref('')
-  const timeout = ref<number | null | undefined>(-1)
-  const color = ref<string | null | undefined>(null)
+  const timeout = ref<number | undefined>(-1)
+  const color = ref<string | undefined>()
   const show = ref(false)
 
   const showMessage = (payload: SnackbarOptions) => {
